chore: rename propellerd to cli
